Include most important Known Issues in readme
First off: thanks for creating and maintaining this really nice tool :)

The readme is great, but could be a bit better if it was a bit more explicit about its current limitations. I think many who try to get their hi refresh rate monitor working will first skim or `cmd+f` the readme for anything related to `refresh` or `144hz` or similar. So this PR adds that text to the readme. 

Thanks!